ai_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch==2.3.0
torchvision==0.18.0
faiss-gpu==1.7.4
transformers==4.42.0
accelerate==0.30.0
bitsandbytes==0.43.1
sentencepiece==0.2.0
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
pymysql==1.1.1
requests==2.32.3
pydantic==2.7.1
